pyproject.toml
[tool.isort]
multi_line_output = 3
profile = "black"
line_length = 79

[tool.mypy]
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pylint]
disable = ["redefined-outer-name"]
max-args = 8
output-format = "colorized"

[tool.poetry]
name = "cytomine-sam"
version = "0.0.1"
description = ""
authors = ["Thibaud Vanmechelen <thibaud.vanmechelen@student.uliege.be>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/ThibaudVanmechelen/Cytomine-sam"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.2"
pillow = "^11.0.0"
pydantic-settings = "^2.8.1"
redis = "^5.2.1"
apscheduler = "^3.11.0"
cytomine-python-client = "^3.0.0"
fastapi = "^0.115.11"
matplotlib = "^3.10.1"
opencv-python = "^4.11.0.86"
pydantic = "^2.10.6"
python-box = "^7.3.2"
sam2 = "^1.1.0"
tomli = "^2.2.1"
torch = {version = "2.6.0+cu118", source = "torch-gpu"}
torchvision = {version = "^0.21.0+cu118", source = "torch-gpu"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
codespell = "^2.2.6"
coverage = "^7.3.2"
isort = "^5.12.0"
pylint = "^3.0.2"
mypy = {extras = ["reports"], version = "^1.7.0"}

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
dvc = {extras = ["gdrive"], version = "^3.29.0"}

[[tool.poetry.source]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"